requirements.txt
emoji
python-dotenv
psycopg2-binary
dostoevsky
nltk
textblob
tortoise-orm
asyncpg
pymorphy
profanity_filter